BUG: Added userpackages importable folder to track down a bug
autowraptest is the folder to try to import into vistrails.
The bug is this: when you import autowraptest, it successfully
creates a module called 'func_wrap_smoke_test' in the
'autowraptest/wraptest' module on the left side of the VisTrails
canvas.  When you add the 'func_wrap_smoke_test' to the canvas,
look at the Inputs in the Module Information on the right
side of the canvas.  Try to expand one of the input ports and a
TypeError is thrown.  Full stack trace will be in issue on
VisTrails gh page.